Invert underlying TLS direction for STREAM ACCEPT
Workaround for https://tahoe-lafs.org/trac/tahoe-lafs/ticket/2861 - as both
sides are underpinned by client endpoints (connecting from app to SAM API), both
sides tried to start the TLS connection. Now, if we detect that the underlying
transport supports TLS, we invert it on the STREAM ACCEPT side, so that from the
outside the transport behaves like a server.

This is not particularly robust; it assumes that the underlying transport always
behaves like a client. It would be better if we could inform the underlying
transport as to whether it should behave like a client or server. See
https://twistedmatrix.com/trac/ticket/3278